Face Recognition
The project included the detection, and recognition of faces stored in
the database, prior to which registration is necessary. The development
software used is visual studio and the code is written in C#. EmguCV
(openCV in .NET) library is used for detection, feature extraction and
recognition of faces, which includes HAAR Casacade algorithm and PCA
algorithm.